added unit test for VoteRequestHandler

+++ src/main/java/org/tools4j/hoverraft/message/VoteRequest.java
@@ -39,6 +39,6 @@ public interface VoteRequest extends Message {
 
     long lastLogIndex();
 
-    VoteRequest lastLogIndex(int lastLogIndex);
+    VoteRequest lastLogIndex(long lastLogIndex);
 
 }
